change default node size to 2xl
I also update the desired capacity to 3 -- everything can fit in 2, but
the system definitely does not run as smoothly.  (Basically I was
getting a lot of alarms.)  When set to 3, there is enough spare room
that everything works well.

It's going to be a bit more expensive, but well worth it IMHO.